Refactor gamemode and termination parsers to be in the termination and gamemode class respectively
